bunch of fix

+++ src/main/java/farn/pacman_arcade/game/Splash.java
@@ -26,11 +26,11 @@ public class Splash {
 
     public void render(Minecraft mc, Screen gui) {
         if (this.tick != 50) {
-            GL11.glEnable((int)3042);
-            GL11.glColor4f((float)1.0f, (float)1.0f, (float)1.0f, (float)1.0f);
-            GL11.glBindTexture((int)3553, (int)mc.textureManager.getTextureId(this.location[this.tick / (50 / this.location.length)]));
+            GL11.glEnable(3042);
+            GL11.glColor4f(1.0f, 1.0f, 1.0f, 1.0f);
+            GL11.glBindTexture(3553, mc.textureManager.getTextureId(this.location[this.tick / (50 / this.location.length)]));
             gui.drawTexture(gui.width / 2 - 84, gui.height / 2 - 100, 0, 0, 168, 200);
-            GL11.glDisable((int)3042);
+            GL11.glDisable(3042);
         } else {
             this.done = true;
         }
